requirements.txt
schedule
PyYAML
apprise